server/src/scripts/rmbg_requirements.txt
modelscope>=1.18.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.39.0,<5.0.0
pillow>=10.0.0
numpy>=1.24.0
kornia>=0.7.0
timm>=0.9.16
